minor fixes, more docs and better organisation

+++ uninstall.sh
@@ -1,15 +1,24 @@
 #!/bin/sh
-# Installer for Waypoints
+
+HELP_TEXT='
+Hello! This is an uninstaller for UnixWaypoints.
+It will:
+1) Erase loading waypoints from shell rc.
+2) Clear all created waypoints (Optional)
+
+Flags:
+-h|--help                   - print this text
+-o|--old                    - use if you had run "instal.sh" from older versions
+-k|--keep-waypoints         - keep my waypoints untouched
+'
 
 WAYPOINTS_FILE="$PWD/.waypoints"
 SCRIPT_FILE="$PWD/waypoints.sh"
 OLD_DIR="$HOME/.local/share/waypoints"
 OLD="$OLD_DIR/waypoints.sh"
-UNINS_FILE="$PWD/uninstall.sh"
-INS_FILE="$PWD/install.sh"
-README_MD="$PWD/README.md"
+MY_MARK="#UnixWaypoints"
 
-# Detect shell startup file
+
 case "$SHELL" in
     */zsh)
         SHELL_RC="$HOME/.zshrc"
@@ -21,19 +30,28 @@ case "$SHELL" in
         SHELL_RC="$HOME/.profile"
         ;;
 esac
-
-grep -vxF ". $SCRIPT_FILE" "$SHELL_RC" > .tmp
+DELETE_WAYPOINTS=1
 case "$1" in
-    --old)
+    -h|--help)
+        echo "$HELP_TEXT"
+        return 1
+        ;;
+    -o|--old)
+        grep -vxF ". $SCRIPT_FILE" "$SHELL_RC" > .tmp
         grep -vxF ". $OLD" .tmp > "$SHELL_RC"
         rm -rf "$OLD_DIR"
         ;;
-    *)
-        cat .tmp > "$SHELL_RC"
+    -k|--keep-waypoints)
+        DELETE_WAYPOINTS=0
         ;;
 esac
+
+NEW_VERSION_EXPORT=$(grep -A 1 "$MY_MARK" "$SHELL_RC")
+grep -vxF "$NEW_VERSION_EXPORT" "$SHELL_RC" > .tmp && cat .tmp > "$SHELL_RC"
+
 rm .tmp
-rm -f "$WAYPOINTS_FILE"
+
+[ $DELETE_WAYPOINTS -eq 1 ] && rm -f "$WAYPOINTS_FILE"
 
 echo "waypoints have been uninstalled"
